fix: providers router (#182)
